Introduce deserialize_datetime_as_string feature flag
When using the Deserializer implementation of this crate
with a serde::de::Deserialize implementation of another crate,
you don't want e.g. 2021-04-03 to become

{"$__toml_private_datetime" = "2021-04-03"}

(which is the special encoding this crate uses to represent
datetimes since serde's data model lacks a type for datetimes)

This commit introduces an off by default feature flag to deserialize
datetimes as strings if so desired, making it feasible to use
the toml crate with other Deserialize implementations without
leaking the above internal representation.